requirements.txt
flask==2.3.3
flask-socketio==5.3.6
requests==2.31.0
python-dotenv==1.0.1
eventlet==0.33.3python-dotenv==1.1.1
